refactor: P2-T15 - 收敛PracticeSessionViewModel并删除业务extensions
