--- LonelyPianistAVP/ViewModels/PracticeSession/PracticeSessionViewModel+ManualReplay.swift
@@ -1,13 +1,0 @@
-import Foundation
-
-extension PracticeSessionViewModel {
-    func startManualReplay(with plan: ManualReplayPlan) {
-        stopVirtualPianoInput()
-        manualReplayCoordinator?.startManualReplay(with: plan)
-    }
-
-    func stopManualReplayTask(restoreAudioRecognition: Bool = true) {
-        manualReplayCoordinator?.stopManualReplayTask(restoreAudioRecognition: restoreAudioRecognition)
-    }
-}
-
