+++ LonelyPianistAVPTests/PracticeHighlightGuideControllerTests.swift
@@ -71,10 +71,16 @@ func transitionGuideSchedulesDelayedSwitchToTrigger() async {
     controller.updateHighlightGuideAfterStepAdvance(previousTick: 0, nextStepIndex: 1)
 
     #expect(store.currentHighlightGuideIndex == 1)
+    for _ in 0..<20 {
+        if await sleeper.recordedDurations().contains(.seconds(0.12)) {
+            break
+        }
+        await Task.yield()
+    }
     #expect(await sleeper.recordedDurations().contains(.seconds(0.12)))
 
     await sleeper.resumeOldest()
-    await Task.yield()
+    for _ in 0..<20 { await Task.yield() }
 
     #expect(store.currentHighlightGuideIndex == 2)
 }
@@ -133,4 +139,3 @@ func shutdownCancelsPendingTransitionTask() async {
 
     #expect(store.currentHighlightGuideIndex == 1)
 }
-
